apis/telemsg/requirements.txt
flask
gunicorn
python-telegram-bot>=12.0.0b1
